+++ libs/react/src/components/application/routes/routes.tsx
@@ -10,7 +10,8 @@ import {
   DocumentUpload,
   Footprint,
   Interceptor,
-  MaterialDetail, MaterialRoutes,
+  MaterialDetail,
+  MaterialRoutes,
   MaterialsPage,
   ProtectedRoute,
   Signup,
@@ -23,26 +24,33 @@ import {
 import { useFlags } from 'launchdarkly-react-client-sdk';
 import { QuestionnaireRoutes } from './questionnaireRoutes';
 
+const DEFAULT_PAGE = '/questionnaires';
+
 export const ColdRoutes = () => {
   const ldFlags = useFlags();
 
   const getFilteredRoutes = () => {
+
     return (
       <>
-        <Route path={'/'} element={<Navigate to={'/compliance'} replace={true} />} />
+        <Route path={'/'} element={<Navigate to={DEFAULT_PAGE} replace={true} />} />
         {ComplianceRoutes()}
         {QuestionnaireRoutes()}
         {ldFlags.showActions261 && <Route path={'/actions'} element={<ActionsOverview />} />}
-        <Route path={'/reports/carbon_footprint'} element={ldFlags.showNewCarbonFootprintModuleCold634 ? <CarbonFootprint /> : <Footprint />} />
+        <Route path={'/carbon_footprint'} element={ldFlags.showNewCarbonFootprintModuleCold634 ? <CarbonFootprint /> : <Footprint />} />
         <Route path={'/documents'} element={ldFlags.showNewDocumentsPage ? <DocumentsPage /> : <DocumentUpload />} />
         <Route path={'/settings/account'} element={<AccountSettingsPage />} />
         <Route path={'/settings/users'} element={<UserSettingsPage />} />
-        <Route path="*" element={<Navigate to={'/compliance'} replace={true} />} />
+        <Route path="*" element={<Navigate to={DEFAULT_PAGE} replace={true} />} />
         {WizardRoutes()}
-        <Route path={'/suppliers'} element={ldFlags.showSuppliersPageCold890 ? <SuppliersPage /> : <Navigate to={'/compliance'} replace={true} />} />
-        <Route path={'/suppliers/:id'} element={ldFlags.showSuppliersPageCold890 ? <SupplierDetail /> : <Navigate to={'/compliance'} replace={true} />} />
+        <Route path={'/suppliers'} element={ldFlags.showSuppliersPageCold890 ? <SuppliersPage /> : <Navigate to={DEFAULT_PAGE} replace={true} />} />
+        <Route path={'/suppliers/:id'} element={ldFlags.showSuppliersPageCold890 ? <SupplierDetail /> : <Navigate to={DEFAULT_PAGE} replace={true} />} />
         {MaterialRoutes()}
-        <Route path={'/settings/billing'} element={ldFlags.showBillingPageCold957 ? <BillingPage /> : <Navigate to={'/compliance'} replace={true} />} />
+        <Route path={'/settings/billing'} element={ldFlags.showBillingPageCold957 ? <BillingPage /> : <Navigate to={DEFAULT_PAGE} replace={true} />} />
+
+        // Temporary redirects from old route until we're certain that the seeds are updated to the new sidebar.
+        <Route path={'/reports/carbon_footprint'} element={<Navigate to={'/carbon_footprint'} replace={true} />} />
+        <Route path={'/compliance'} element={<Navigate to={'/questionnaires'} replace={true} />} />
       </>
     );
   };